Re-factor in preparation for focal stat map op
 - Rename TerrainIndexMapOp to RawFocalMapOp and move to mrgeo-core
 - Modify RawFocalMapOp so that it processes all bands
 - for issue #186
