+++ mrgeo-mapalgebra/mrgeo-mapalgebra-terrain/src/main/scala/org/mrgeo/mapalgebra/TpiMapOp.scala
@@ -2,8 +2,9 @@ package org.mrgeo.mapalgebra
 
 import java.awt.image.{WritableRaster, Raster}
 
+import org.mrgeo.image.MrsPyramidMetadata
 import org.mrgeo.mapalgebra.parser.{ParserException, ParserNode}
-import org.mrgeo.mapalgebra.raster.RasterMapOp
+import org.mrgeo.mapalgebra.raster.{RawFocalMapOp, RasterMapOp}
 
 object TpiMapOp extends MapOpRegistrar {
   def create(raster: RasterMapOp): MapOp =
@@ -17,23 +18,23 @@ object TpiMapOp extends MapOpRegistrar {
     new TpiMapOp(node, variables)
 }
 
-class TpiMapOp extends TerrainIndexMapOp
+class TpiMapOp extends RawFocalMapOp
 {
   // Chad suggested a 33 pixel "radius" by default for TPI for good results
-  private var kernelSize: Int = 67
+  private var neighborhoodSize: Int = 67
 
   private[mapalgebra] def this(node:ParserNode, variables: String => Option[ParserNode]) = {
     this()
 
     if (node.getNumChildren != 1 && node.getNumChildren != 2) {
-      throw new ParserException("tpi usage: tpi(rasterInput, [kernel size])")
+      throw new ParserException("tpi usage: tpi(rasterInput, [neighborhood size])")
     }
     inputMapOp = RasterMapOp.decodeToRaster(node.getChild(0), variables)
     if (node.getNumChildren == 2) {
-      val kernelArg = MapOp.decodeInt(node.getChild(1), variables)
-      kernelArg match {
-        case Some(k) => kernelSize = k
-        case None => throw new ParserException("Expected a number for the kernel size")
+      val neighborhoodArg = MapOp.decodeInt(node.getChild(1), variables)
+      neighborhoodArg match {
+        case Some(k) => neighborhoodSize = k
+        case None => throw new ParserException("Expected a number for the neighborhood size")
       }
     }
   }
@@ -46,13 +47,13 @@ class TpiMapOp extends TerrainIndexMapOp
   override protected def computePixelValue(rasterValues: Array[Double], notnodata: Array[Boolean],
                                            rasterWidth: Int,
                                            processX: Int, processY: Int,
-                                           xLeftOffset: Int, kernelWidth: Int,
-                                           yAboveOffset: Int, kernelHeight: Int, tileId: Long): Double =
+                                           xLeftOffset: Int, neighborhoodWidth: Int,
+                                           yAboveOffset: Int, neighborhoodHeight: Int, tileId: Long): Double =
   {
     var x: Int = processX - xLeftOffset
-    val maxX = x + kernelWidth
+    val maxX = x + neighborhoodWidth
     var y: Int = processY - yAboveOffset
-    val maxY = y + kernelHeight
+    val maxY = y + neighborhoodHeight
     val processPixel: Double = rasterValues(calculateRasterIndex(rasterWidth, processX, processY))
     var sum: Double = 0.0
     var count: Int = 0
@@ -73,13 +74,13 @@ class TpiMapOp extends TerrainIndexMapOp
   }
 
   /**
-    * Returns 2 values about the kernel to use (kernel width, kernel height).
+    * Returns 2 values about the neighborhood to use (neighborhood width, neighborhood height).
     *
     * This method is called at the start of the execution of this map op.
     *
     * @return
     */
-  override protected def getKernelInfo: (Int, Int) = {
-    (kernelSize, kernelSize)
+  override protected def getNeighborhoodInfo: (Int, Int) = {
+    (neighborhoodSize, neighborhoodSize)
   }
 }
